py/requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.1
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
opencensus-proto==0.1.0
opentelemetry-api==0.6b0
opentelemetry-auto-instrumentation==0.6b0
opentelemetry-ext-http-requests==0.6b0
opentelemetry-ext-otcollector==0.6b0
opentelemetry-ext-wsgi==0.6b0
opentelemetry-sdk==0.6b0
protobuf==3.11.3
pylint==2.4.4
requests==2.23.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.1
wrapt==1.11.2
